Fix confusingly-translated questions?
The outsourcing question at line 492 is currently unclear regarding whether the quiz-taker supports the claim that it's evil or the claim that it's necessary. Given the axis it modifies, though, it's clearly about the claim that outsourcing is necessary.

The minimum-penalty question at line 1084 is currently unclear regarding whether the quiz-taker thinks light punishments are unjust or whether the quiz taker thinks laws saying 'you must punish a criminal at least this much' are unjust. Given the axis it modifies, though, it's clearly about mandatory minimum sentencing.

The questions at lines 1204 and 1425 are merely phrased non-idiomatically.